requirements.txt
Flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
google-generativeai==0.8.3
protobuf==5.28.2
gunicorn==22.0.0  # For Render deployment